chore(lib): move lib/swc in lib
