+++ .github/workflows/update-swc.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Check if SWC update is required
         id: version-check
         run: |
-          CURRENT_SWC_VERSION=$(cat lib/swc/package.json | jq -r '.version')
+          CURRENT_SWC_VERSION=$(cat lib/package.json | jq -r '.version')
           if [[ -n "${{ github.event.inputs.swc_version }}" ]]; then
             NEW_SWC_VERSION="${{ github.event.inputs.swc_version }}"
           else
