+++ lib/wasm.d.ts
