+++ lib/wasm.js
